WIP base character refactoring
Preparing for character abilities and simplifying implementation of classes. Added Master class since we need the ability to assign a class to a character, not just making a completely new class based on their chosen class.

Current issues: not calling update on all player components. only working with Warrior class. Having the player component scripts as variables in a master script seems to cause the Update() method on the inner scripts not to be called, so we need to call all their updates in master's update if we are doing it this way